Int to Long

+++ src/main/java/com/transferwise/acorn/models/Step.java
@@ -4,9 +4,9 @@ import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
-public class Step{
-    public int id;
-    public int transactionId;
+public class Step {
+    public Long id;
+    public Long transactionId;
     public String type;
     public Date creationTime;
     public List<BalancesAfter> balancesAfter;
@@ -15,8 +15,8 @@ public class Step{
     public Object tracingReferenceCode;
     public SourceAmount sourceAmount;
     public TargetAmount targetAmount;
-    public int sourceBalanceId;
-    public int targetBalanceId;
+    public Long sourceBalanceId;
+    public Long targetBalanceId;
     public Fee fee;
     public BigDecimal rate;
 }
